Remove #if defined tags which span over multiple lines
I had a problem when trying to parse headers with `if defined` inside,
that were not defined during the compilation. The solution, which is not
really ideal, was to strip away anything within this tags. As I do not
need anything between that tags at the moment it solved my problem but I
am already thinking if the connection to unity and its compiler defines
could be used somehow?

I have added the unit test for my example.