Added array_as_byte config options.
[Anders Esbensen]

Option to check arrays by the number of bytes instead of number of elements.

[Philippe Coval]

This change was applied to a fork (ths_cmock) used in a Z-Wave project.
It was needed to generate mocks from a header that used annonymous structures.

The observed issue was:

(...)_mock.c:486:123: error: invalid application of ‘sizeof’ to incomplete type ‘struct (...)’

Which was just declared in relative header.

The array_as_byte feature prevent this error.

Author: Anders Esbensen <Anders.Esbensen@silabs.com>
Origin-Upstream: https://github.com/ThrowTheSwitch/CMock/pull/503
Origin-SiliconLabs: v2.5.1-74-ga078d10
Relate-to: https://github.com/SiliconLabsSoftware/z-wave-protocol-controller/issues/75
Signed-off-by: Philippe Coval <philippe.coval@silabs.com>